docs: simple JSON/FITS result loading entry point in guides/results (#65)
* docs: document loading results from output folder (.json/.fits)

Restructure guides/results/ to present two distinct result-loading paths:
the new top-level start_here.py covers direct .json/.fits loading from the
output folder (fast, in-memory), while the former start_here.py is moved
into a renamed aggregator/ folder to document the memory-efficient
generator-based workflow. Both entry points cross-link to explain the
trade-off. modeling.py gains a __Loading From Output Folder__ section
with a tracer.json / tracer.fits example pointing to the new guide.


* docs: use search.paths.output_path for portable result loading example


* fix: use autoconf.from_json for result loading and guard on file existence

Tracer.from_json / Galaxies.from_json are not valid APIs — objects are
deserialised via autoconf.dictable.from_json. Also guard the modeling.py
example with Path.exists() so it no-ops under PYAUTO_SKIP_FIT_OUTPUT=1
(used in smoke tests) while still running after a normal fit.


---------
