+++ scripts/guides/results/aggregator/samples_via_aggregator.py
@@ -2,7 +2,7 @@
 Results: Samples via Aggregator
 ===============================
 
-In the script `autogalaxy_workspace/*/guides/results/examples/samples.py` we show how to inspect the non-linear
+In the script `autogalaxy_workspace/*/guides/results/aggregator/samples.py` we show how to inspect the non-linear
 search samples from a result.
 
 We have also shown how to use the `Aggregator` to load the samples of a non-linear search from hard-disk or a
@@ -41,7 +41,7 @@ __Contents__
 __Samples via Result__
 
 A fraction of this example repeats the API for manipulating samples given in the
-`autogalaxy_workspace/*/guides/results/examples/samples.py` example.
+`autogalaxy_workspace/*/guides/results/aggregator/samples.py` example.
 
 This is done so users can directly copy and paste Python code which loads results from the database and manipulates
 the samples.
