+++ scripts/guides/results/database/start_here.py
@@ -346,6 +346,6 @@ __Wrap Up__
 
 This example illustrates how to use the database.
 
-The API above can be combined with the `results/examples` scripts in order to use the database to load results and
+The API above can be combined with the `results/aggregator` scripts in order to use the database to load results and
 perform analysis on them.
 """
