+++ scripts/guides/units/cosmology.py
@@ -20,7 +20,7 @@ __Contents__
 __Errors__
 
 To produce errors on unit converted quantities, you`ll may need to perform marginalization over samples of these
-converted quantities (see `results/examples/samples.ipynb`).
+converted quantities (see `results/aggregator/samples.ipynb`).
 """
 
 from autoconf import jax_wrapper  # Sets JAX environment before other imports
